Fix klas-setup.sql: cast row_number naar int in klas_week streak
Postgres heeft geen operator date - bigint; row_number() is bigint.
De klasstreak-berekening (expected-datumreeks) castte niet, waardoor
de hele SQL afbrak op 'operator does not exist: date - bigint'.
Cast row_number()-uitdrukking naar ::int (date - int is wél geldig).

+++ sql/klas-setup.sql
@@ -362,7 +362,7 @@ returns jsonb language sql stable security definer set search_path = public as $
                 else null end as a
   ),
   seq as (
-    select d, (select a from anchor) - (row_number() over (order by d desc) - 1) as expected
+    select d, (select a from anchor) - (row_number() over (order by d desc) - 1)::int as expected
       from days where d <= (select a from anchor)
   )
   select jsonb_build_object(
